Moved setting buttons to their page
- Added a "Launch Team Fortress 2" button.
- Moved the download source list to the top of the screen.
