+++ src/TF2HUD.Editor/SettingsWindow.xaml.cs
@@ -1,4 +1,7 @@
-﻿using System.Windows;
+﻿using System.Globalization;
+using System.Windows;
+using HUDEditor.Properties;
+using WPFLocalizeExtension.Engine;
 
 namespace HUDEditor
 {
@@ -10,6 +13,43 @@ namespace HUDEditor
         public SettingsWindow()
         {
             InitializeComponent();
+
+            // Check for app updates.
+            BtnAutoUpdate.IsChecked = Settings.Default.app_update_auto;
+        }
+
+        /// <summary>
+        ///     Updates localization to the selected language.
+        /// </summary>
+        private void BtnLocalize_OnClick(object sender, RoutedEventArgs e)
+        {
+            if (BtnLocalizeEn.IsChecked == true)
+                LocalizeDictionary.Instance.Culture = new CultureInfo("en-US");
+            else if (BtnLocalizeFr.IsChecked == true)
+                LocalizeDictionary.Instance.Culture = new CultureInfo("fr-FR");
+            else if (BtnLocalizeRu.IsChecked == true)
+                LocalizeDictionary.Instance.Culture = new CultureInfo("ru-RU");
+
+            // Save language preference to user settings.
+            Settings.Default.user_language = LocalizeDictionary.Instance.Culture.ToString();
+            Settings.Default.Save();
+        }
+
+        private void BtnSetDirectory_OnClick(object sender, RoutedEventArgs e)
+        {
+            MainWindow.Logger.Info("Attempting to change the 'tf/custom' directory.");
+            MainWindow.SetupDirectory(true);
+        }
+
+        private void BtnRefresh_OnClick(object sender, RoutedEventArgs e)
+        {
+            MainWindow.CheckSchemaUpdates(false);
+        }
+
+        private void BtnAutoUpdate_OnClick(object sender, RoutedEventArgs e)
+        {
+            Settings.Default.app_update_auto = BtnAutoUpdate.IsChecked ?? true;
+            Settings.Default.Save();
         }
     }
 }
